human_loop_platform/requirements.txt
# Core Dependencies
streamlit==1.29.0
pandas==2.1.3
numpy==1.26.2
plotly==5.18.0

# AI & ML
google-generativeai>=0.8.0
openai==1.3.7
scikit-learn==1.3.2
xgboost==2.0.2

# File Processing
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1
pillow==10.1.0
python-magic==0.4.27
sqlparse==0.4.4

# Backend & API
fastapi==0.104.1
uvicorn==0.24.0
celery==5.3.4
redis==5.0.1
pydantic==2.5.2
python-multipart==0.0.6

# Notebook Integration
marimo==0.1.88
jupyter==1.0.0
nbformat==5.9.2

# Database & Storage
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
minio==7.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
playwright==1.40.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
aiofiles==23.2.1
watchdog==3.0.0